Now the code can be compiled again, and the executable works. Time to start to design everything from the beginning.

+++ mytypes.hpp
@@ -1,3 +1,3 @@
 
 //Definition of some types
-typedef long mlong;
+typedef long long mlong;
